+++ src/archives.rs
@@ -2,21 +2,25 @@ use crate::error::Error;
 use flate2::read::GzDecoder;
 use std::fs::{self, File};
 use std::path::Path;
-use tar::Archive;
 use tempfile::tempdir_in;
+use zip_extensions::read::zip_extract;
 
 /// Supported archive types.
 pub(crate) enum ArchiveFormat {
     TarGz,
+    Zip,
 }
 
 impl ArchiveFormat {
     /// Parse archive type from resource extension.
     pub(crate) fn parse_from_extension(resource: &str) -> Result<Self, Error> {
         if resource.ends_with(".tar.gz") {
-            return Ok(Self::TarGz);
+            Ok(Self::TarGz)
+        } else if resource.ends_with(".zip") {
+            Ok(Self::Zip)
+        } else {
+            Err(Error::ExtractionError("unsupported archive format".into()))
         }
-        Err(Error::ExtractionError("unsupported archive format".into()))
     }
 }
 
@@ -33,8 +37,15 @@ pub(crate) fn extract_archive<P: AsRef<Path>>(
         ArchiveFormat::TarGz => {
             let tar_gz = File::open(path)?;
             let tar = GzDecoder::new(tar_gz);
-            let mut archive = Archive::new(tar);
+            let mut archive = tar::Archive::new(tar);
             archive.unpack(&temp_target)?;
+        }
+        ArchiveFormat::Zip => {
+            zip_extract(
+                &path.as_ref().to_path_buf(),
+                &temp_target.path().to_path_buf(),
+            )
+            .map_err(|e| Error::ExtractionError(format!("{:?}", e)))?;
         }
     };
